requirements.txt
playwright==1.50.0
playwright-stealth==1.0.6
2captcha-python==1.5.0
pyyaml==6.0.2
requests==2.32.3
